+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -226,7 +226,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01JZAR8H5A0HGMG4C7K1Z58QEC"
+        "01JZMR03EGA03VCJ8PX3ZNZVQJ"
     )
     r"""The id of the resource"""