+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasetdatapointsop.py
@@ -600,7 +600,7 @@ class ListDatasetDatapointsData(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-07-04T12:58:16.632Z")
+    updated: Optional[datetime] = parse_datetime("2025-07-08T10:06:05.308Z")
     r"""The date and time the resource was last updated"""