+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -125,7 +125,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01JZAR8H534GM7NWZXNZ24QXTN"
+        "01JZMR03EECKYXJD2H1RNS5MS2"
     )
     r"""The id of the resource"""