+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -100,7 +100,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01JZAR8H5BXK4NBGH6FWGXWC54"
+        "01JZMR03EH4T9K72F9EB2BBE7A"
     )
     r"""The id of the resource"""